# api/openapi.yaml
openapi: 3.1.0
info:
  title: KakEnBetaal API
  version: 0.1.0
  description: HTTP contract. Money is integer cents. Keep in sync with code; types are generated with `pnpm gen:api`.
servers:
  - url: http://localhost:3000
...
components:
  securitySchemes:
    bearer: {type: http, scheme: bearer, bearerFormat: JWT}
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties: {code: {type: string}, message: {type: string}, details: {type: object}}
...
  responses:
    Err: {description: Error, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
security: []
paths:
...
  /t/{tagCode}:
    get:
      tags: [public]
      summary: Tag tap. Verifies tag (SDM if NTAG424), resolves bill, 302 to wa.me with one-time token.
      parameters:
        - {name: tagCode, in: path, required: true, schema: {type: string}}
        - {name: picc_data, in: query, schema: {type: string}}
        - {name: cmac, in: query, schema: {type: string}}
      responses:
        '302': {description: Redirect to WhatsApp}
        '404': {$ref: '#/components/responses/Err'}
        '403': {$ref: '#/components/responses/Err'}
  /b/{billToken}:
    get:
      tags: [public]
      summary: Bill link or QR. Mints a one-time claim token bound to this bill and redirects to WhatsApp. Anyone holding the link may pay.
      parameters: [{name: billToken, in: path, required: true, schema: {type: string}}]
      responses: {'302': {description: Redirect}, '404': {$ref: '#/components/responses/Err'}}
  /r/{receiptToken}:
    get:
      tags: [public]
      summary: Receipt page (slip with print animation, sound toggle, PDF link). Private, no-store, strict CSP. 404 for unknown or revoked tokens.
      parameters: [{name: receiptToken, in: path, required: true, schema: {type: string, pattern: '^[A-Za-z0-9_-]{32}$'}}]
      responses: {'200': {description: Receipt, content: {text/html: {schema: {type: string}}}}, '404': {$ref: '#/components/responses/Err'}}
  /r/{receiptToken}/slip.png:
    get:
      tags: [public]
      summary: Slip image (PNG, under 1 MB) for a receipt token
      parameters: [{name: receiptToken, in: path, required: true, schema: {type: string}}]
      responses: {'200': {description: PNG, content: {image/png: {schema: {type: string, format: binary}}}}, '404': {$ref: '#/components/responses/Err'}}
  /r/{receiptToken}/slip.pdf:
    get:
      tags: [public]
      summary: The slip as a PDF (download)
      parameters: [{name: receiptToken, in: path, required: true, schema: {type: string}}]
      responses: {'200': {description: PDF, content: {application/pdf: {schema: {type: string, format: binary}}}}, '404': {$ref: '#/components/responses/Err'}}
  /i/{invoiceToken}:
    get:
      tags: [public]
      summary: An issued tax invoice (PDF). The customer gets the link on WhatsApp after replying INVOICE.
      parameters: [{name: invoiceToken, in: path, required: true, schema: {type: string, pattern: '^[A-Za-z0-9_-]{32}$'}}]
      responses: {'200': {description: PDF, content: {application/pdf: {schema: {type: string, format: binary}}}}, '404': {$ref: '#/components/responses/Err'}}
...
  /sim/outbox:
    get:
      tags: [public]
      summary: Dev only (WA_MODE=sim). Messages the bot sent to one number, for the WhatsApp simulator.
      parameters:
        - {name: to, in: query, required: true, schema: {type: string}}
        - {name: after, in: query, schema: {type: integer, minimum: 0}}
      responses: {'200': {description: messages}}
  /webhooks/whatsapp:
    get: {tags: [public], summary: Meta verification handshake, responses: {'200': {description: challenge echoed}}}
    post:
      tags: [public]
      summary: Inbound WhatsApp events. Requires valid X-Hub-Signature-256.
      responses: {'200': {description: accepted}, '401': {$ref: '#/components/responses/Err'}}
  /webhooks/provider/{provider}:
    post:
      tags: [public]
      summary: Provider payment events. Signature verified by adapter. Idempotent.
      parameters: [{name: provider, in: path, required: true, schema: {type: string, enum: [mock, peach, payfast]}}]
      responses: {'200': {description: accepted}, '401': {$ref: '#/components/responses/Err'}}
  /mock-checkout/{ref}:
    get: {tags: [public], summary: Mock hosted checkout (dev only), parameters: [{name: ref, in: path, required: true, schema: {type: string}}], responses: {'200': {description: page}}}
    post: {tags: [public], summary: Resolve mock checkout, parameters: [{name: ref, in: path, required: true, schema: {type: string}}], responses: {'200': {description: ok}}}
...